+++ sample/src/main/java/com/clevertap/demo/ui/main/HomeScreenFragment.kt
@@ -57,7 +57,9 @@ class HomeScreenFragment : Fragment() {
         listItemBinding = HomeScreenFragmentBinding.inflate(layoutInflater, container, false).apply {
             viewmodel = viewModel
         }
-        listItemBinding.expandableListView.isNestedScrollingEnabled = true
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
+            listItemBinding.expandableListView.isNestedScrollingEnabled = true
+        }
 
         return listItemBinding.root
     }